Removed and replaced the XML window with an ImGui one (#4)
* Removed and replaced the XML window with an ImGui one

* Removed the XML in game window.
* Added ImGui Window (same as the settings panel)
* Reworked the /pluginman command to toggle ImGui Window open | closed.
* Tossed a toggle button on the settings panel to toggle the main window.

* removed the toggle button

removed the toggle button and HelpMarker and replaced with simple Text explaining /pluginman command

* addresses comments

* Added MQ2PluginManager.ini setting to save window state.
* Window will show at load the first time.
* Setting saves when display state changes.
* Removed game state check

* dropped the version number from initialize output.

the version number really did nothing as it was just a string output on initialize.

* save state when manually closing window

this way the state is saved both manually closing or using the toggle command.

the prior was only saving the state with the slash command.

* readded the game state check

* removed unneeded header

* addresses toggeling out of game

Does not automatically show out of game you have to toggle it.

opening and closing the window both in and out of game will save that state. 

Will automatically appear in game based on saved state.